+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [2025.8.1] - 2025-08-01
+
 ### Added
 
 - Ability to retry when a request has failed.
@@ -449,7 +451,8 @@ Initial release with the following features:
 - services like opening the gate,
 - user attendance history.
 
-[unreleased]: https://github.com/coworking-metz/mobile-app/compare/2025.7.7...main
+[unreleased]: https://github.com/coworking-metz/mobile-app/compare/2025.8.1...main
+[2025.8.1]: https://github.com/coworking-metz/mobile-app/compare/2025.7.7...2025.8.1
 [2025.7.7]: https://github.com/coworking-metz/mobile-app/compare/2025.7.6...2025.7.7
 [2025.7.6]: https://github.com/coworking-metz/mobile-app/compare/2025.7.5...2025.7.6
 [2025.7.5]: https://github.com/coworking-metz/mobile-app/compare/2025.7.4...2025.7.5